Related work items: #820, #829

+++ client/src/hooks/chat/useRoomMenuDialog.js
@@ -5,6 +5,7 @@ import ApiContext from "../../context/ApiContext";
 import * as actionChat from "../../store/actions/chat";
 
 import dialogs, { ROOM_MENU_DIALOG_TYPE } from "../../constants/dialogs";
+import { BLOCKED_TYPE } from "../../constants/block_types";
 
 import useDialog from "../useDialog";
 import useBlockDialog from "../../hooks/useBlockDialog";
@@ -14,9 +15,11 @@ export default function useRoomMenuDialog(callback) {
   const dispatch = useDispatch();
   const apiClient = useContext(ApiContext);
 
-  const blockDialog = useBlockDialog(async ({ userName, postId }) => {
-    await dispatch(actionChat.removeRoom(postId));
-    actionChat.resetCurrentRoom();
+  const blockDialog = useBlockDialog(async ({ userName, postId, blockType }) => {
+    if (blockType === BLOCKED_TYPE) {
+      await dispatch(actionChat.removeRoom(postId));
+      dispatch(actionChat.resetCurrentRoom());
+    }
   });
 
   const handleDeleteRoom = useCallback(async (id) => {
